Extract nav bar to a single file
