+++ mobile/.vscode/launch.json
@@ -7,7 +7,8 @@
         {
             "name": "mobile",
             "request": "launch",
-            "type": "dart"
+            "type": "dart",
+            "program": "lib/app/main.dart"
         }
     ]
 }
