+++ mobile/lib/features/common/nav_bar.dart
@@ -1,0 +1,68 @@
+import 'package:flutter/material.dart';
+import 'package:mobile/resources/app_colors.dart';
+
+/// This is the stateful widget that the main application instantiates.
+class NavBar extends StatefulWidget {
+  NavBar({Key key}) : super(key: key);
+
+  @override
+  _NavBarState createState() => _NavBarState();
+}
+
+/// This is the private State class that goes with NavBar.
+class _NavBarState extends State<NavBar> {
+  int _selectedIndex = 0;
+  static const TextStyle optionStyle =
+      TextStyle(fontSize: 30, fontWeight: FontWeight.bold);
+  static const List<Widget> _widgetOptions = <Widget>[
+    Text(
+      'Index 0: Home',
+      style: optionStyle,
+    ),
+    Text(
+      'Index 1: Family',
+      style: optionStyle,
+    ),
+    Text(
+      'Index 2: My Profile',
+      style: optionStyle,
+    ),
+  ];
+
+  void _onItemTapped(int index) {
+    setState(() {
+      _selectedIndex = index;
+    });
+  }
+
+  @override
+  Widget build(BuildContext context) {
+    return Scaffold(
+      appBar: AppBar(
+        title: const Text('Welcome!'),
+      ),
+      body: Center(
+        child: _widgetOptions.elementAt(_selectedIndex),
+      ),
+      bottomNavigationBar: BottomNavigationBar(
+        items: const <BottomNavigationBarItem>[
+          BottomNavigationBarItem(
+            icon: Icon(Icons.home),
+            label: 'Home',
+          ),
+          BottomNavigationBarItem(
+            icon: Icon(Icons.nature_people),
+            label: 'Family',
+          ),
+          BottomNavigationBarItem(
+            icon: Icon(Icons.account_circle),
+            label: 'My Profile',
+          ),
+        ],
+        currentIndex: _selectedIndex,
+        selectedItemColor: AppColors.main,
+        onTap: _onItemTapped,
+      ),
+    );
+  }
+}
